# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hospital Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transfers:
    get:
      operationId: listTransfers
      tags: [transfers]
      summary: List all hospital transfers
      responses:
        "200":
          description: List of transfers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transfer"
    post:
      operationId: createTransfer
      tags: [transfers]
      summary: Create a hospital transfer request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransferRequest"
      responses:
        "201":
          description: Transfer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /transfers/{id}:
    get:
      operationId: getTransfer
      tags: [transfers]
      summary: Get a transfer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transfer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /transfers/{id}/accept:
    post:
      operationId: acceptTransfer
      tags: [transfers]
      summary: Accept an incoming transfer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transfer accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /transfers/{id}/reject:
    post:
      operationId: rejectTransfer
      tags: [transfers]
      summary: Reject a transfer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Transfer rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /transfers/{id}/in-transit:
    post:
      operationId: markTransferInTransit
      tags: [transfers]
      summary: Mark transfer as in-transit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transfer in-transit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /transfers/{id}/arrived:
    post:
      operationId: markTransferArrived
      tags: [transfers]
      summary: Mark transfer as arrived
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transfer arrived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /transfers/{id}/cancel:
    post:
      operationId: cancelTransfer
      tags: [transfers]
      summary: Cancel a transfer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transfer cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
...
components:
  schemas:
...
    TransferStatus:
      type: string
      enum: [PENDING, ACCEPTED, IN_TRANSIT, ARRIVED, REJECTED, CANCELLED]
...
    Hospital:
      type: object
      properties:
        id:
          type: integer
        nameEn:
          type: string
        nameAr:
          type: string
        city:
          type: string
        address:
          type: string
        phone:
          type: string
        specialties:
          type: array
          items:
            type: string
        isAvailable:
          type: string
        level:
          type: string
      required:
        - id
        - nameEn
        - nameAr
        - city
        - address
        - phone
        - specialties
        - isAvailable
        - level
    CreateTransferRequest:
      type: object
      properties:
        caseId:
          type: integer
        direction:
          type: string
          enum: [OUTGOING, INCOMING]
        toHospitalId:
          type: integer
        fromHospitalId:
          type: integer
        patientName:
          type: string
        patientAge:
          type: integer
        patientGender:
          type: string
        chiefComplaint:
          type: string
        riskLevel:
          type: string
        transportMethod:
          type: string
          enum: [AMBULANCE, AIR_AMBULANCE, PRIVATE_VEHICLE, HOSPITAL_TRANSPORT]
        clinicalSummary:
          type: string
        specialRequirements:
          type: string
        estimatedArrival:
          type: string
          format: date-time
      required:
        - direction
        - patientName
        - patientAge
        - patientGender
        - chiefComplaint
        - riskLevel
        - transportMethod
        - clinicalSummary
    Transfer:
      type: object
      properties:
        id:
          type: integer
        caseId:
          type: integer
        direction:
          type: string
        fromHospitalId:
          type: integer
        toHospitalId:
          type: integer
        fromHospital:
          $ref: "#/components/schemas/Hospital"
        toHospital:
          $ref: "#/components/schemas/Hospital"
        patientName:
          type: string
        patientAge:
          type: integer
        patientGender:
          type: string
        chiefComplaint:
          type: string
        riskLevel:
          type: string
        transportMethod:
          type: string
        clinicalSummary:
          type: string
        specialRequirements:
          type: string
        estimatedArrival:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/TransferStatus"
        rejectionReason:
          type: string
        transferCode:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - direction
        - fromHospitalId
        - toHospitalId
        - patientName
        - patientAge
        - patientGender
        - chiefComplaint
        - riskLevel
        - transportMethod
        - clinicalSummary
        - status
        - transferCode
        - createdAt
        - updatedAt
...