# lib/api-zod/src/generated/types/createTransferRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hospital Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTransferRequestDirection } from "./createTransferRequestDirection";
import type { CreateTransferRequestTransportMethod } from "./createTransferRequestTransportMethod";

export interface CreateTransferRequest {
  caseId?: number;
  direction: CreateTransferRequestDirection;
  toHospitalId?: number;
  fromHospitalId?: number;
  patientName: string;
  patientAge: number;
  patientGender: string;
  chiefComplaint: string;
  riskLevel: string;
  transportMethod: CreateTransferRequestTransportMethod;
  clinicalSummary: string;
  specialRequirements?: string;
  estimatedArrival?: Date;
}