# lib/api-zod/src/generated/types/hospital.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hospital Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Hospital {
  id: number;
  nameEn: string;
  nameAr: string;
  city: string;
  address: string;
  phone: string;
  specialties: string[];
  isAvailable: string;
  level: string;
}