# lib/api-zod/src/generated/types/transfer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hospital Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Hospital } from "./hospital";
import type { TransferStatus } from "./transferStatus";

export interface Transfer {
  id: number;
  caseId?: number;
  direction: string;
  fromHospitalId: number;
  toHospitalId: number;
  fromHospital?: Hospital;
  toHospital?: Hospital;
  patientName: string;
  patientAge: number;
  patientGender: string;
  chiefComplaint: string;
  riskLevel: string;
  transportMethod: string;
  clinicalSummary: string;
  specialRequirements?: string;
  estimatedArrival?: Date;
  status: TransferStatus;
  rejectionReason?: string;
  transferCode: string;
  createdAt: Date;
  updatedAt: Date;
}