# lib/api-zod/src/generated/types/transferStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Hospital Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export type TransferStatus =
  (typeof TransferStatus)[keyof typeof TransferStatus];

export const TransferStatus = {
  PENDING: "PENDING",
  ACCEPTED: "ACCEPTED",
  IN_TRANSIT: "IN_TRANSIT",
  ARRIVED: "ARRIVED",
  REJECTED: "REJECTED",
  CANCELLED: "CANCELLED",
} as const;